+++ lib/karafka/backends/sidekiq.rb
@@ -5,7 +5,7 @@ module Karafka
     # Sidekiq backend that schedules stuff to Sidekiq worker for delayed execution
     module Sidekiq
       # Karafka Sidekiq backend version
-      VERSION = '1.2.0.beta2'
+      VERSION = '1.2.0.beta3'
 
       # Enqueues the execution of perform method into a worker.
       # @note Each worker needs to have a class #perform_async method that will allow us to pass
@@ -16,7 +16,7 @@ module Karafka
         Karafka.monitor.instrument('backends.sidekiq.process', caller: self) do
           topic.worker.perform_async(
             topic.id,
-            topic.interchanger.load(params_batch)
+            topic.interchanger.encode(params_batch)
           )
         end
       end