+++ spec/lib/karafka/interchanger_spec.rb
@@ -20,36 +20,11 @@ RSpec.describe Karafka::Interchanger do
     )
   end
 
-  describe '#load' do
-    it 'expect to return parsed version of itself' do
-      expect(interchanger_class.load(params_batch)).to eq parsed_params_batch
-    end
-  end
-
-  describe '#parse' do
-    let(:params_batch) do
-      [
-        {
-          'value' => 1,
-          'parser' => Class,
-          'receive_time' => 1,
-          'a' => 1
-        }
-      ]
-    end
-    let(:parsed_params_batch) do
-      [
-        {
-          'a' => 1,
-          parser: Class,
-          value: 1,
-          receive_time: 1
-        }
-      ]
-    end
-
-    it 'expect to return what was provided with symbolized system keys' do
-      expect(interchanger_class.parse(params_batch)).to eq parsed_params_batch
+  describe 'encode and decode chain' do
+    it 'expects to be able to encode and decode and remain structure' do
+      encoded = described_class.encode(params_batch)
+      decoded = described_class.decode(encoded)
+      expect(decoded).to eq parsed_params_batch
     end
   end
 end