+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Karafka Sidekiq Backend
 
-## 1.2.0.beta2
+## 1.2.0.beta3
+- ```#load``` and ```#parse``` are renamed to ```#encode``` and ```#decode``` in interchangers
+- Default interchanger is not Base64 Marshal dump
 - #274 - Rename controllers to consumers
 - Karafka 1.2 support
 - Parse data before pushing to Sidekiq